fix: simplify project dependencies, fix tests
Many of the explicitly referenced dependencies are actually a transitive dependency of some of the others. Especially with test dependencies, where some transitive dependencies were excluded in favor of yet another explicit ones. This was quite hard to untangle and if there ever was a trouble with those versions, it's likely not the case with 8.7.1 and above.

I was also quite confused with the 'plugin.version' property, which seemed to mimic the implicit 'project.version'. These were also likely intermixed in plugin-descriptor.properties, so this change tries to fix that by replacing 'plugin.version' with 'project.version' where appropriate, ultimately removing the 'plugin.version'. Same was done with 'lucene.version' which was there only to draw in an explicit version of 'org.apache.lucene:lucene-test-framework', despite the correct version already being a dependency of 'org.elasticsearch.test:framework'.

Despite all the simplifications above the tests still had problems executing. This was ultimately resolved by using a newer version of the maven-surefire-plugin.